softfloat: clean bfloat16 type usage

+++ models/cpu/iss/softfloat/f16_to_ui32.cpp
@@ -88,7 +88,7 @@ uint_fast32_t f16_to_ui32_alt( float16_t a, uint_fast8_t roundingMode, bool exac
     union ui16_f16 uA;
     uint_fast16_t uiA;
     bool sign;
-    int_fast8_t exp;
+    int_fast16_t exp;
     uint_fast32_t sig;
     uint_fast64_t sig64;
     int_fast16_t shiftDist;
